requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
black==24.8.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.4.0
distlib==0.3.8
dnspython==2.6.1
filelock==3.16.1
frozenlist==1.4.1
identify==2.6.1
idna==3.10
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==3.8.0
pycares==4.4.0
pycparser==2.22
pymongo==4.9.1
PyYAML==6.0.2
requests==2.32.3
termcolor==2.4.0
urllib3==2.2.3
virtualenv==20.26.5
yarl==1.11.1
